services/fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
prometheus-client==0.20.0
mlflow==2.16.0
scikit-learn==1.5.1
pandas==2.2.2
boto3==1.34.162
